Added lib function for getting current user session
